+++ db/migrate/20191201014634_create_lesson_table_rows.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class CreateLessonTableRows < ActiveRecord::Migration[5.2]
+  def change
+    create_view :lesson_table_rows
+  end
+end